Line up lib directory with `src` directory
Previously a change was committed/merged to the Coffeescript version of the
code but not to the Javascript version.

Adds a Makefile and some convenience commands for checking that the code runs
as anticipated. Updates Coffeescript to 1.8.0 (it's the version that was used
to compile the most recent Javascript) and stores the command used to compile
it in a Makefile.

Updates Travis CI to use the Makefile instead of the Cakefile, and to use the
new container-based infrastructure.